Move taiko skinning tests to own namespace
